+++ extra/group/include/rmvl/group/rune_group.h
@@ -55,9 +55,9 @@ public:
      * - 计算整个序列组当前帧的原始数据 `raw_data`，并同步至 `datas` 中
      *
      * @param[in] gyro_data 最新陀螺仪数据
-     * @param[in] tick 最新时间戳
+     * @param[in] tick 最新时间点
      */
-    void sync(const GyroData &gyro_data, int64_t tick) override;
+    void sync(const GyroData &gyro_data, double tick) override;
 
     /**
      * @brief 获取原始数据队列
